Add files via upload
This project is a Simple Parking Management System with CRUD. This is a web-based application 
project developed in Python, using with using Django frame work, HTML, CSS, JS, and MySQL 
Database. The main goal of this application is to provide Vehicle Parking Businesses with an 
automated and online platform to manage their daily transactions. This can help the said type of 
business to easily store, retrieve, and manage their daily transactions with their customers. The 
application has a pleasant user interface using the Bootstrap 4.6 Framework and consists of user
friendly features and functionalities which can give the end-users a better experience while using 
the app.